Server renamed to Hub. Make Hub return the IChannel on connection, delegating the send/receive methods/events to the application layer.
