+++ Yacs/Exceptions/HubFailureException.cs
@@ -3,16 +3,16 @@
 namespace Yacs.Exceptions
 {
     /// <summary>
-    /// Represents an error in the <see cref="Server"/>.
+    /// Represents an error in the <see cref="Hub"/>.
     /// </summary>
-    public class ServerFailureException : Exception
+    public class HubFailureException : Exception
     {
         /// <summary>
-        /// Creates a new <see cref="ServerFailureException"/>.
+        /// Creates a new <see cref="HubFailureException"/>.
         /// </summary>
         /// <param name="message">A message indicating the cause of the exception.</param>
         /// <param name="innerException">The exception that is the cause of the current exception.</param>
-        internal ServerFailureException(string message, Exception innerException) : base(message, innerException)
+        internal HubFailureException(string message, Exception innerException) : base(message, innerException)
         {
         }
     }
