--- Yacs/Options/ServerOptions.cs
@@ -1,38 +1,0 @@
-﻿namespace Yacs.Options
-{
-    /// <summary>
-    /// Represents the options to create a <see cref="Server"/>.
-    /// </summary>
-    public class ServerOptions : BaseOptions
-    {
-        /// <summary>
-        /// Gets if a server is discoverable. Default: yes 
-        /// </summary>
-        public bool IsDiscoverable => DiscoveryPort > 0;
-
-        /// <summary>
-        /// Gets or sets the port used for the discovery. Default: 11000
-        /// </summary>
-        public int DiscoveryPort { get; set; }
-
-        /// <summary>
-        /// Gets or sets the maximum number of clients allowed. Default: Infinite
-        /// </summary>
-        public int MaximumChannels { get; set; }
-
-        /// <summary>
-        /// Gets or sets if the channels in the server should be monitored for disconnections. Default: false
-        /// </summary>
-        public bool ActiveChannelMonitoring { get; set; }
-
-        /// <summary>
-        /// Creates a new <see cref="ServerOptions"/> object with the default values.
-        /// </summary>
-        public ServerOptions() : base()
-        {
-            DiscoveryPort = 11000;
-            MaximumChannels = 0;
-            ActiveChannelMonitoring = false;
-        }
-    }
-}
