+++ Yacs/Services/OptionsValidator.cs
@@ -10,16 +10,13 @@ namespace Yacs.Services
 
         private const int MAXIMUM_CHANNELS_LOWER_LIMIT = 0;
 
-        public static void Validate(ServerOptions serverOptions)
+        public static void Validate(HubOptions serverOptions)
         {
             if (serverOptions.DiscoveryPort < DISCOVERY_PORT_LOWER_LIMIT)
-                throw new OptionsException($"{nameof(ServerOptions)}.{nameof(ServerOptions.DiscoveryPort)} must be equal to or greater than {DISCOVERY_PORT_LOWER_LIMIT}.");
+                throw new OptionsException($"{nameof(HubOptions)}.{nameof(HubOptions.DiscoveryPort)} must be equal to or greater than {DISCOVERY_PORT_LOWER_LIMIT}.");
 
             if (serverOptions.DiscoveryPort > DISCOVERY_PORT_UPPER_LIMIT)
-                throw new OptionsException($"{nameof(ServerOptions)}.{nameof(ServerOptions.DiscoveryPort)} must be less than or equal to {DISCOVERY_PORT_UPPER_LIMIT}.");
-
-            if (serverOptions.MaximumChannels < MAXIMUM_CHANNELS_LOWER_LIMIT)
-                throw new OptionsException($"{nameof(ServerOptions)}.{nameof(ServerOptions.MaximumChannels)} must be equal to or greater than {MAXIMUM_CHANNELS_LOWER_LIMIT}.");
+                throw new OptionsException($"{nameof(HubOptions)}.{nameof(HubOptions.DiscoveryPort)} must be less than or equal to {DISCOVERY_PORT_UPPER_LIMIT}.");
 
             Validate(serverOptions as BaseOptions);
         }
@@ -35,7 +32,7 @@ namespace Yacs.Services
             {
                 var maximumBytesForOneEncodedCharacter = baseOptions.Encoding.GetMaxByteCount(1);
                 if (baseOptions.ReceptionBufferSize < maximumBytesForOneEncodedCharacter)
-                    throw new OptionsException($"{nameof(ServerOptions)}.{nameof(ServerOptions.ReceptionBufferSize)} must be equal to or greater than the maximum number of bytes that can be produced by encoding a single {baseOptions.Encoding.HeaderName} character ({maximumBytesForOneEncodedCharacter}).");
+                    throw new OptionsException($"{nameof(HubOptions)}.{nameof(HubOptions.ReceptionBufferSize)} must be equal to or greater than the maximum number of bytes that can be produced by encoding a single {baseOptions.Encoding.HeaderName} character ({maximumBytesForOneEncodedCharacter}).");
             }
         }
     }
